#291: refactored into Util.js and mcdev-copado.d.js
